requirements.txt
anyio==3.7.1
boto3==1.33.13
botocore==1.33.13
certifi==2025.6.15
charset-normalizer==3.4.2
exceptiongroup==1.3.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.10
jmespath==1.0.1
python-dateutil==2.9.0.post0
python-telegram-bot==20.3
requests==2.31.0
s3transfer==0.8.2
six==1.17.0
sniffio==1.3.1
typing-extensions==4.7.1
urllib3==1.26.20
